+++ R/data-mask.R
@@ -103,7 +103,7 @@ DataMask <- R6Class("DataMask",
     },
 
     set_current_group = function(group) {
-      parent.env(private$chops)$.current_group <- group
+      parent.env(private$chops)$.current_group[] <- group
     },
 
     full_data = function() {
@@ -192,6 +192,10 @@ DataMask <- R6Class("DataMask",
 
     get_rlang_mask = function() {
       private$mask
+    },
+
+    get_caller_env = function() {
+      private$caller
     }
 
   ),