+++ NEWS.md
@@ -1,3 +1,18 @@
+# dplyr 1.0.5
+
+* Fixed edge case of `slice_sample()` when `weight_by=` is used and there 
+  0 rows (#5729). 
+  
+* `across()` can again use columns in functions defined inline (#5734). 
+
+* Using testthat 3rd edition. 
+
+* Fixed bugs introduced in `across()` in previous version (#5765).
+
+* `group_by()` keeps attributes unrelated to the grouping (#5760).
+
+* The `.cols=` argument of `if_any()` and `if_all()` defaults to `everything()`. 
+
 # dplyr 1.0.4
 
 * Improved performance for `across()`. This makes `summarise(across())` and